requirements.txt
langchain-groq
langchain
langchain_community
chainlit
nbconvert
python-dotenv
numexpr
wikipedia
groq

